pytest_demo/pytest_server_test/requirements.txt
pytest>=7.4.0
allure-pytest>=2.13.0
requests>=2.31.0
paramiko>=3.3.0
pyyaml>=6.0
loguru>=0.7.0